refactor : usepoint() 주석 제거 및 형변환 일치

+++ src/main/java/com/sparta/devcamp_spring/payment/entity/Point.java
@@ -24,7 +24,7 @@ public class Point extends BaseEntity {
     @OneToMany(mappedBy = "point")
     private List<PointLog> logs;
 
-    public void use(int amountToUse) {
+    public void use(Double amountToUse) {
         this.availableAmount -= amountToUse;
     }
